Create README.md
a project designed to study the trends of deforestation in the world from 1990 to 2016. 
It aims at answering many questions regarding deforestation and the collective actions that countries must take in tackling the problem. The data used in this project is obtained from the world bank, comprising of the world forest area and land area from 1990 to 2016.